Added most of collisions(with oranges and with butter). Code needs so much refactor......Jeez!

+++ micro-machines-solution/Vector3.h
@@ -13,7 +13,7 @@ public:
 	double getX();
 	double getY();
 	double getZ();
-	int norm();
+	double norm();
 	void set(double, double, double);
 	Vector3 operator=(const Vector3&);
 	Vector3 operator+(const Vector3&);
